Reorganize membership CRUD operations
