+++ driven/storage/adapter.go
@@ -1018,182 +1018,6 @@ func (sa *Adapter) FindUserGroups(clientID string, userID string, category *stri
 	return list, nil
 }
 
-// CreatePendingMember creates a pending member for a specific group
-func (sa *Adapter) CreatePendingMember(clientID string, user *model.User, group *model.Group, member *model.Member) error {
-	if member != nil && group != nil {
-
-		//1. check if the user is already a member of this group - pending or member or admin or rejected
-		membership, err := sa.FindGroupMembership(clientID, group.ID, user.ID)
-		if err == nil && membership != nil {
-			switch membership.Status {
-			case "admin":
-				return errors.New("the user is an admin for the group")
-			case "member":
-				return errors.New("the user is a member for the group")
-			case "pending":
-				return errors.New("the user is pending for the group")
-			case "rejected":
-				return errors.New("the user is rejected for the group")
-			default:
-				return errors.New("error creating a pending user")
-			}
-		}
-
-		//2. check if the answers match the group questions
-		if len(group.MembershipQuestions) != len(member.MemberAnswers) {
-			return errors.New("member answers mismatch")
-		}
-
-		member.ID = uuid.NewString()
-		member.DateCreated = time.Now().UTC()
-
-		_, err = sa.db.groupMemberships.InsertOne(member.ToGroupMembership(clientID, group.ID))
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// DeletePendingMember deletes a pending member from a specific group
-func (sa *Adapter) DeletePendingMember(clientID string, groupID string, userID string) error {
-	return sa.DeleteMember(clientID, groupID, userID, true)
-}
-
-// CreateMemberUnchecked Created a member to a group
-func (sa *Adapter) CreateMemberUnchecked(clientID string, current *model.User, group *model.Group, member *model.Member) error {
-	if group != nil {
-		membership, err := sa.FindGroupMembership(clientID, group.ID, current.ID)
-		if err != nil || membership == nil || !membership.IsAdmin() {
-			log.Printf("error: storage.CreateMemberUnchecked() - current user is not admin of the group")
-			return fmt.Errorf("current user is not admin of the group")
-		}
-
-		existingMembership, _ := sa.FindGroupMembership(clientID, group.ID, member.UserID)
-		if existingMembership != nil {
-			log.Printf("error: storage.CreateMemberUnchecked() - member of group '%s' with user id %s already exists", group.Title, member.UserID)
-			return fmt.Errorf("member of group '%s' with user id %s already exists", group.Title, member.UserID)
-		}
-
-		existingMembership, _ = sa.FindGroupMembership(clientID, group.ID, member.ExternalID)
-		if existingMembership != nil {
-			log.Printf("error: storage.CreateMemberUnchecked() - member of group '%s' with external id %s already exists", group.Title, member.ExternalID)
-			return fmt.Errorf("member of group '%s' with external id %s already exists", group.Title, member.ExternalID)
-		}
-
-		if len(member.UserID) == 0 && len(member.ExternalID) == 0 {
-			log.Printf("error: storage.CreateMemberUnchecked() - expected user_id or external_id")
-			return fmt.Errorf("expected user_id or external_id")
-		}
-
-		member.ID = uuid.NewString()
-		member.DateCreated = time.Now()
-		member.MemberAnswers = group.CreateMembershipEmptyAnswers()
-
-		_, err = sa.db.groupMemberships.InsertOne(member.ToGroupMembership(clientID, group.ID))
-		if err != nil {
-			return err
-		}
-
-	}
-
-	return nil
-}
-
-// DeleteMember deletes a member membership from a specific group
-func (sa *Adapter) DeleteMember(clientID string, groupID string, userID string, force bool) error {
-
-	err := sa.db.dbClient.UseSession(context.Background(), func(sessionContext mongo.SessionContext) error {
-		err := sessionContext.StartTransaction()
-		if err != nil {
-			log.Printf("error starting a transaction - %s", err)
-			return err
-		}
-
-		currentMembership, _ := sa.FindGroupMembershipWithContext(sessionContext, clientID, groupID, userID)
-		if currentMembership != nil {
-
-			if currentMembership.IsAdmin() {
-				adminMemberships, _ := sa.FindGroupMembershipsWithContext(sessionContext, clientID, model.MembershipFilter{
-					GroupIDs: []string{groupID},
-					Statuses: []string{"admin"},
-				})
-				if len(adminMemberships.Items) <= 1 && !force {
-					log.Printf("sa.DeleteMember() - there must be at least two admins in order to delete ")
-					return fmt.Errorf("there must be at least two admins in order to delete ")
-				}
-			}
-
-			filter := bson.D{
-				primitive.E{Key: "group_id", Value: groupID},
-				primitive.E{Key: "user_id", Value: userID},
-				primitive.E{Key: "client_id", Value: clientID},
-			}
-			_, err := sa.db.groupMemberships.DeleteOneWithContext(sessionContext, filter, nil)
-			if err != nil {
-				abortTransaction(sessionContext)
-				log.Printf("error deleting membership - %s", err)
-				return err
-			}
-		}
-
-		err = sessionContext.CommitTransaction(sessionContext)
-		if err != nil {
-			fmt.Println(err)
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// ApplyMembershipApproval applies a membership approval
-func (sa *Adapter) ApplyMembershipApproval(clientID string, membershipID string, approve bool, rejectReason string) error {
-
-	status := "rejected"
-	if approve {
-		status = "member"
-	}
-
-	filter := bson.D{primitive.E{Key: "_id", Value: membershipID}, primitive.E{Key: "client_id", Value: clientID}}
-	update := bson.D{
-		primitive.E{Key: "$set", Value: bson.D{
-			primitive.E{Key: "status", Value: status},
-			primitive.E{Key: "reject_reason", Value: rejectReason},
-			primitive.E{Key: "date_updated", Value: time.Now()},
-		},
-		},
-	}
-	_, err := sa.db.groupMemberships.UpdateOne(filter, update, nil)
-	return err
-}
-
-// DeleteMembership deletes a membership
-func (sa *Adapter) DeleteMembership(clientID string, current *model.User, membershipID string) error {
-	filter := bson.D{primitive.E{Key: "_id", Value: membershipID}, primitive.E{Key: "client_id", Value: clientID}}
-	_, err := sa.db.groupMemberships.DeleteOne(filter, nil)
-	return err
-}
-
-// UpdateMembership updates a membership
-func (sa *Adapter) UpdateMembership(clientID string, current *model.User, membershipID string, status string, dateAttended *time.Time) error {
-	filter := bson.D{primitive.E{Key: "_id", Value: membershipID}, primitive.E{Key: "client_id", Value: clientID}}
-	update := bson.D{
-		primitive.E{Key: "$set", Value: bson.D{
-			primitive.E{Key: "status", Value: status},
-			primitive.E{Key: "date_attended", Value: dateAttended},
-			primitive.E{Key: "date_updated", Value: time.Now()},
-		},
-		},
-	}
-	_, err := sa.db.groupMemberships.UpdateOne(filter, update, nil)
-	return err
-}
-
 // FindEvents finds the events for a group
 func (sa *Adapter) FindEvents(clientID string, current *model.User, groupID string, filterByToMembers bool) ([]model.Event, error) {
 	filter := bson.D{
